+++ ft_memmove.c
@@ -6,11 +6,13 @@
 /*   By: tjorge-l <tjorge-l@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 12:47:21 by tjorge-l          #+#    #+#             */
-/*   Updated: 2024/04/15 14:29:08 by tjorge-l         ###   ########.fr       */
+/*   Updated: 2024/04/17 15:35:53 by tjorge-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *dest, const void *src, unsigned int n)
+#include "libft.h"
+
+void	*ft_memmove(void *dest, const void *src, size_t n)
 {
 	char	*dest_pt;
 	char	*src_pt;
